+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 10.2.0 (2019-12-25)
 
 * `stubbles\environments\errorhandler\LogErrorHandler` now logs request uri as additional field if present
+* `stubbles\environments\exceptionhandler\AbstractExceptionHandler` now logs request uri as additional field if present
 * `stubbles\ExceptionLogger::log()` now accepts the request id as optional second argument
 
 ## 10.1.0 (2019-12-18)